% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age_structure}
\alias{age_structure}
\title{Set-up age groups and find ageing rate
Uses the global variables numMonthly and final_age, along
with the appropriate greater Perth data file.
Assumes a non-uniform, but constant population distribution}
\usage{
age_structure(year)
}
\arguments{
\item{year}{ABS greater Perth population year}
}
\value{
list nAges, pop_groups, age_in, age_out
}
\description{
Set-up age groups and find ageing rate
Uses the global variables numMonthly and final_age, along
with the appropriate greater Perth data file.
Assumes a non-uniform, but constant population distribution
}
