% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_loglikelihood}
\alias{calc_loglikelihood}
\title{Calculate log likelihood}
\usage{
calc_loglikelihood(modelOut, data, risk = 1)
}
\arguments{
\item{modelOut}{model output from aggregate_output function}

\item{data}{observational incidence data}

\item{risk}{1 if risk model, 0 for base}
}
\value{
numeric
}
\description{
Calculate log likelihood
}
