% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cohort_ageing}
\alias{cohort_ageing}
\title{Cohort ageing}
\usage{
cohort_ageing(counts, vacc = 0)
}
\arguments{
\item{counts}{table of compartment counts}

\item{vacc}{if vacc model 1, else 0}
}
\value{
list
}
\description{
Cohort ageing
}
