% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{comp_data}
\alias{comp_data}
\title{Comparison plot of 2 data sets over same time range}
\usage{
comp_data(
  obs1,
  obs2,
  ylabel = "Hospitalisations",
  years = 2000:2013,
  text = "RSV hospitalisations: observed (colour .) and model (black --)"
)
}
\arguments{
\item{obs1}{table of monthly hospitalisation data in age groups (colored dots)}

\item{obs2}{table of monthly hospitalisation data in age groups (dashed line)}

\item{ylabel}{default is Hospitalisations}

\item{years}{year labels for x-axis}

\item{text}{plot text, default is RSV hospitalisations: observed (colour .) and model (black --)}
}
\value{
ggplot
}
\description{
Comparison plot of 2 data sets over same time range
}
