% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{comp_series}
\alias{comp_series}
\title{Plot of 2 time series of hospitalisation or infection data for 1 age group}
\usage{
comp_series(
  obs1,
  obs2,
  ylabel = "Hospitalisations children < 5yo",
  years = 2000:2013,
  text = ""
)
}
\arguments{
\item{obs1}{vector of monthly hospitalisation data (colored dots)}

\item{obs2}{vector of monthly hospitalisation data (dashed line)}

\item{ylabel}{default is Hospitalisations}

\item{years}{year labels for x-axis}

\item{text}{plot text, default is RSV hospitalisations: observed (colour .) and model (black --)}
}
\value{
ggplot
}
\description{
Plot of 2 time series of hospitalisation or infection data for 1 age group
}
