% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_likelihood_risk}
\alias{create_likelihood_risk}
\title{Set up likelihood function for risk model in a closure environment
for use with lazymcmc}
\usage{
create_likelihood_risk(parTab, data, fixedPar, PRIOR_FUNC, ...)
}
\arguments{
\item{parTab}{parameter table setup for use in lazymcmc fitting functions}

\item{data}{observational incidence data}

\item{fixedPar}{fixed parameter values}

\item{PRIOR_FUNC}{optional prior function}
}
\value{
function
}
\description{
Set up likelihood function for risk model in a closure environment
for use with lazymcmc
}
