% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.greaterPerth.2011}
\alias{data.greaterPerth.2011}
\title{ABS population of Greater Perth by 5-year age group - 2011}
\format{
A data table containing 18 observations of 4 variables.
\describe{
\item{lga}{Local Government Area classification}
\item{lower.age.limit}{}
\item{year}{census data year}
\item{population}{population in age group}
}
}
\source{
\url{https://abs.gov.au/census/find-census-data/quickstats/2021/5GPER}
}
\usage{
data.greaterPerth.2011
}
\description{
Data is formatted to be easily used with the R package conmat
to find the applicable contact matrix
}
\keyword{datasets}
