% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{deSolve_base}
\alias{deSolve_base}
\title{Base ODE model with multiple exposures}
\usage{
deSolve_base(t, y, parms)
}
\arguments{
\item{t}{a vector of times at which model output is wanted}

\item{y}{a matrix of the initial values of the state variables.
Each column corresponds to a state variable.}

\item{parms}{the required model parameters and their values. For this model,
parms <- pars_pT <- list(b0 = b0,
b1 = b1,
phi = phi,
delta = delta,
gamma0 = gamma0,
gamma1 = gamma1,
nu = nu,
omega_vect = omega_vect,
omegaE = omegaE,
A = A,
B = B,
C = C,
D = D,
age_in = age_in,
age_out = age_out,
age_months = age_months,
sigma_vect = sigma_vect,
sigmaE = sigmaE,
mixing = mixing,
nAges = nAges)}
}
\value{
list
}
\description{
This function is used with R package deSolve to solve the ODEs
}
