% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findAverageAgePerMonth}
\alias{findAverageAgePerMonth}
\title{Find average age of first infection/hospitalisation, second+
infection/hospitalisation for all, term and preterm
NOTE: Only coded for risk model}
\usage{
findAverageAgePerMonth(
  modOut,
  ageYears,
  ageRange = 1/12,
  startT = NA,
  endT = NA,
  risk = 1,
  obs = 0
)
}
\arguments{
\item{modOut}{model output returned from aggregate_output function or
observed hospitalisations in table form}

\item{ageYears}{lower bound of age groups in years}

\item{ageRange}{age range of each age group, default is 1 month}

\item{startT}{start month, default NA for observed hospitalisations}

\item{endT}{end month, default NA for observed hospitalisations}

\item{risk}{if 1, then risk model, 0 for base #not coded for 0 yet}

\item{obs}{if 1, then finding average age in observed hospitalisations
(combined terms and preterms)}
}
\value{
list
}
\description{
Find average age of first infection/hospitalisation, second+
infection/hospitalisation for all, term and preterm
NOTE: Only coded for risk model
}
