% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findIndexRange}
\alias{findIndexRange}
\title{From model output, find the time range that matches to the observed
data set. Do this by assuming a "peak" month, default is 7 (July)}
\usage{
findIndexRange(counts, month = 7, obs)
}
\arguments{
\item{counts}{list of modelled detected infections in aggregated age groups,
in same form as observational data}

\item{month}{month 1 - 12, Default is 7 (July)}

\item{obs}{table of monthly hospitalisation data for 4 age groups, at this
stage assuming complete calendar years}
}
\value{
list(index_start, index_end, phase)
}
\description{
From model output, find the time range that matches to the observed
data set. Do this by assuming a "peak" month, default is 7 (July)
}
