% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initialise}
\alias{initialise}
\title{Initialise compartments}
\usage{
initialise(vacc = 0, file = 0)
}
\arguments{
\item{vacc}{if vaccination model 1, else 0}

\item{file}{if 1, use initial values from file init_conds.rds}
}
\value{
list
}
\description{
Initialise compartments
}
