% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{plot_ageToRisk}
\alias{plot_ageToRisk}
\title{Plot age-to-risk function}
\usage{
plot_ageToRisk(A, B, C, D, E, maxAge = 12)
}
\arguments{
\item{A}{the average maximum risk increase for a term infant (at age 0)}

\item{B}{the exponential function decay constant}

\item{C}{the average minimum risk over all ages for all infants}

\item{D}{the scaling factor (<1) modifying risk for the prior infection group}

\item{E}{the scaling factor (>1) modifying risk for those born preterm}

\item{maxAge}{maximum age in months to plot}
}
\value{
ggplot
}
\description{
Plot age-to-risk function
}
