% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_fitToData}
\alias{plot_fitToData}
\title{Compare model(s) predicted hospitalisations to observed hospitalisation data}
\usage{
plot_fitToData(pars1, pars2 = NA, obs, max_t = 200)
}
\arguments{
\item{pars1}{list of parameter values from model 1}

\item{pars2}{list of compartment counts from model 2 if comparing}

\item{obs}{table of monthly hospitalisation data for 4 age groups}

\item{max_t}{the "run-in" time used in fitting the model}
}
\value{
ggplot
}
\description{
Compare model(s) predicted hospitalisations to observed hospitalisation data
}
