% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_fitToData95}
\alias{plot_fitToData95}
\title{Compare model predicted hospitalisations to observed hospitalisation data
with 95\% confidence interval from mcmc derived distributions}
\usage{
plot_fitToData95(parTab, obs, max_t = 200)
}
\arguments{
\item{parTab}{table of values, with following headings:
names, bestPars, lower2.5 and upper97.5}

\item{obs}{table of monthly hospitalisation data for 4 age groups}

\item{max_t}{the "run-in" time used in fitting the model}
}
\value{
ggplot
}
\description{
Compare model predicted hospitalisations to observed hospitalisation data
with 95\% confidence interval from mcmc derived distributions
}
