% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{plot_mod}
\alias{plot_mod}
\title{Plot model output}
\usage{
plot_mod(outT, hos = T, plotText = "")
}
\arguments{
\item{outT}{table of monthly hospitalisation/infection data aggregated
into age groups}

\item{hos}{if T, y axis label is "Hospitalisation" else "Infections". Default is T}

\item{plotText}{title text. Default is no title}
}
\value{
ggplot
}
\description{
Plot model output
}
