% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_rsv_model}
\alias{run_rsv_model}
\title{Run base model}
\usage{
run_rsv_model(
  b0,
  b1,
  phi,
  h_k1,
  h_k2,
  h_k3,
  h_k4,
  max_t = 2000,
  init_conds_from_file = 0,
  save_init_conds = 0
)
}
\arguments{
\item{b1}{amplitude of seasonal forcing (fitted)}

\item{phi}{phase of seasonal forcing (fitted)}

\item{h_k1}{prop. of infections that are detected
in 0-2 month (fitted)}

\item{h_k2}{prop. of infections that are detected
in 3-5 month age group (fitted)}

\item{h_k3}{prop. of infections that are detected
in 6-11 month age group (fitted )}

\item{h_k4}{prop. of infections that are detected
in 12-23 month age group (fitted )}

\item{max_t}{number of model runs}

\item{init_conds_from_file}{1 = read inital values from file,
0 = use default from parameters.r}

\item{save_init_conds}{1 = save resulting model values in rds file to
use as initial values for next iteration
0 = do not save model values to file}

\item{bo}{transmission coeff. (fitted)}
}
\value{
data.frame
}
\description{
Run base model
}
