% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setup_model}
\alias{setup_model}
\title{Set-up model structure and define fixed parameter values as global variables.
Either choose a model version or pass values for parameters and structure.}
\usage{
setup_model(version = 1, uniform = 1, pars)
}
\arguments{
\item{uniform}{if using uniform age distribution 1 else 0}
}
\value{
1
}
\description{
Set-up model structure and define fixed parameter values as global variables.
Either choose a model version or pass values for parameters and structure.
}
