% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{comp_data}
\alias{comp_data}
\title{Comparison plot of 2 data sets over same time range}
\usage{
comp_data(obs1, obs2)
}
\arguments{
\item{obs1}{table of monthly hospitalisation data for 4 age groups}

\item{obs2}{table of monthly hospitalisation data for 4 age groups}
}
\value{
ggplot
}
\description{
Comparison plot of 2 data sets over same time range
}
