% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_contact_matrix}
\alias{get_contact_matrix}
\title{Get monthly contact matrix - will add to for sub-population specific, COVID measures etc.
Note: This produces a symmetric matrix which is different to legacy
contact matrix "mixing_75.csv". mixing_75 is NOT symmetric}
\usage{
get_contact_matrix(type = 0, uniform = 1)
}
\arguments{
\item{type}{0 is legacy matrix used in foundation modeling work
(adjusted for age structure)
1 is polymod Great Britain matrix as retrieved from
R package socialmixr
2 is contact matrix from conmat using polymod adjusted for
2014 greater perth population NOTE: Takes a while to run,
might want to save as R data file}

\item{uniform}{1 uniform population distribution, 0 non-uniform. Only
used with conmat matrix, type 2}
}
\value{
matrix
}
\description{
Get monthly contact matrix - will add to for sub-population specific, COVID measures etc.
Note: This produces a symmetric matrix which is different to legacy
contact matrix "mixing_75.csv". mixing_75 is NOT symmetric
}
